+++ src/Aws/CloudWatchLogs/Resources/cloudwatchlogs-2014-03-28.php
@@ -511,6 +511,91 @@ return array (
                 ),
             ),
         ),
+        'FilterLogEvents' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'FilterLogEventsResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'Logs_20140328.FilterLogEvents',
+                ),
+                'logGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+                'logStreamNames' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'minItems' => 1,
+                    'maxItems' => 100,
+                    'items' => array(
+                        'name' => 'LogStreamName',
+                        'type' => 'string',
+                        'minLength' => 1,
+                        'maxLength' => 512,
+                    ),
+                ),
+                'startTime' => array(
+                    'type' => 'numeric',
+                    'location' => 'json',
+                ),
+                'endTime' => array(
+                    'type' => 'numeric',
+                    'location' => 'json',
+                ),
+                'filterPattern' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                    'maxLength' => 512,
+                ),
+                'nextToken' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                ),
+                'limit' => array(
+                    'type' => 'numeric',
+                    'location' => 'json',
+                    'minimum' => 1,
+                    'maximum' => 10000,
+                ),
+                'interleaved' => array(
+                    'type' => 'boolean',
+                    'format' => 'boolean-string',
+                    'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Returned if a parameter of the request is incorrectly specified.',
+                    'class' => 'InvalidParameterException',
+                ),
+                array(
+                    'reason' => 'Returned if the specified resource does not exist.',
+                    'class' => 'ResourceNotFoundException',
+                ),
+                array(
+                    'reason' => 'Returned if the service cannot complete the request.',
+                    'class' => 'ServiceUnavailableException',
+                ),
+            ),
+        ),
         'GetLogEvents' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -996,6 +1081,57 @@ return array (
                             ),
                             'creationTime' => array(
                                 'type' => 'numeric',
+                            ),
+                        ),
+                    ),
+                ),
+                'nextToken' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+        ),
+        'FilterLogEventsResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'events' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'FilteredLogEvent',
+                        'type' => 'object',
+                        'properties' => array(
+                            'logStreamName' => array(
+                                'type' => 'string',
+                            ),
+                            'timestamp' => array(
+                                'type' => 'numeric',
+                            ),
+                            'message' => array(
+                                'type' => 'string',
+                            ),
+                            'ingestionTime' => array(
+                                'type' => 'numeric',
+                            ),
+                            'eventId' => array(
+                                'type' => 'string',
+                            ),
+                        ),
+                    ),
+                ),
+                'searchedLogStreams' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'SearchedLogStream',
+                        'type' => 'object',
+                        'properties' => array(
+                            'logStreamName' => array(
+                                'type' => 'string',
+                            ),
+                            'searchedCompletely' => array(
+                                'type' => 'boolean',
                             ),
                         ),
                     ),