+++ src/Aws/Route53Domains/Resources/route53domains-2014-05-15.php
@@ -121,6 +121,10 @@ return array (
                     'reason' => 'The number of operations or jobs running exceeded the allowed threshold for the account.',
                     'class' => 'OperationLimitExceededException',
                 ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
+                ),
             ),
         ),
         'DisableDomainAutoRenew' => array(
@@ -155,6 +159,10 @@ return array (
                 array(
                     'reason' => 'The requested item is not acceptable. For example, for an OperationId it may refer to the ID of an operation that is already completed. For a domain name, it may not be a valid domain name or belong to the requester account.',
                     'class' => 'InvalidInputException',
+                ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
                 ),
             ),
         ),
@@ -203,6 +211,10 @@ return array (
                     'reason' => 'The number of operations or jobs running exceeded the allowed threshold for the account.',
                     'class' => 'OperationLimitExceededException',
                 ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
+                ),
             ),
         ),
         'EnableDomainAutoRenew' => array(
@@ -237,6 +249,10 @@ return array (
                 array(
                     'reason' => 'The requested item is not acceptable. For example, for an OperationId it may refer to the ID of an operation that is already completed. For a domain name, it may not be a valid domain name or belong to the requester account.',
                     'class' => 'InvalidInputException',
+                ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
                 ),
             ),
         ),
@@ -285,6 +301,10 @@ return array (
                     'reason' => 'The number of operations or jobs running exceeded the allowed threshold for the account.',
                     'class' => 'OperationLimitExceededException',
                 ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
+                ),
             ),
         ),
         'GetDomainDetail' => array(
@@ -319,6 +339,10 @@ return array (
                 array(
                     'reason' => 'The requested item is not acceptable. For example, for an OperationId it may refer to the ID of an operation that is already completed. For a domain name, it may not be a valid domain name or belong to the requester account.',
                     'class' => 'InvalidInputException',
+                ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
                 ),
             ),
         ),
@@ -471,6 +495,10 @@ return array (
                 array(
                     'reason' => 'The number of operations or jobs running exceeded the allowed threshold for the account.',
                     'class' => 'OperationLimitExceededException',
+                ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
                 ),
             ),
         ),
@@ -818,6 +846,10 @@ return array (
                 array(
                     'reason' => 'The requested item is not acceptable. For example, for an OperationId it may refer to the ID of an operation that is already completed. For a domain name, it may not be a valid domain name or belong to the requester account.',
                     'class' => 'InvalidInputException',
+                ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
                 ),
             ),
         ),
@@ -1428,6 +1460,10 @@ return array (
                     'reason' => 'The number of operations or jobs running exceeded the allowed threshold for the account.',
                     'class' => 'OperationLimitExceededException',
                 ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
+                ),
             ),
         ),
         'UpdateDomainContactPrivacy' => array(
@@ -1489,6 +1525,10 @@ return array (
                 array(
                     'reason' => 'The number of operations or jobs running exceeded the allowed threshold for the account.',
                     'class' => 'OperationLimitExceededException',
+                ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
                 ),
             ),
         ),
@@ -1565,6 +1605,10 @@ return array (
                     'reason' => 'The number of operations or jobs running exceeded the allowed threshold for the account.',
                     'class' => 'OperationLimitExceededException',
                 ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
+                ),
             ),
         ),
         'UpdateTagsForDomain' => array(
@@ -1619,6 +1663,10 @@ return array (
                 array(
                     'reason' => 'The number of operations or jobs running exceeded the allowed threshold for the account.',
                     'class' => 'OperationLimitExceededException',
+                ),
+                array(
+                    'reason' => 'Amazon Route 53 does not support this top-level domain.',
+                    'class' => 'UnsupportedTLDException',
                 ),
             ),
         ),